% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapStatistics.R
\name{gapStatistics}
\alias{gapStatistics}
\title{gap statistics}
\usage{
gapStatistics(d, K = 3, B = 10, gamma = NULL, alpha = 1, group = NULL,
  seed = 15213, silence = FALSE)
}
\arguments{
\item{d}{combined data matrix n*J, where n is number of subjects, J=J1+J2+... and J1 is number of features in omics dataset 1 and J2 is number of features in omics dataset 2...}

\item{K}{number of clusters}

\item{B}{number of permutations.}

\item{gamma}{Penalty on total number of features. Larger gamma will yeild small number of selected features.}

\item{alpha}{balance between group sparsity and individual sparsity. alpha=1 yeilds no group sparsity. alpha=0 yeilds no individual penalty.}

\item{group}{Prior group information. Potentially these group can contain overlap features. group is a list and each element of the list is feature index.}

\item{seed}{random seed}

\item{silence}{do not print progress in silence = TRUE.}
}
\value{
a table. Each row represents a gamma. 
\item{gamma}{input gammas}
\item{score}{objective score: see obj0 in ISKmeans function}
\item{E.score}{mean value of permutated score}
\item{se.score}{standard error of permutated score}
\item{gapStat}{gap statistics, score - E.score}
\item{numF}{number of selected features}
}
\description{
gap statistics to tune gamma
}
\details{
gap statistics to tune gamma, the tuning parameter to control number of features.
}
\examples{
set.seed(123)

# Generate two random omics datasets
mu <- c(-3,1,3)
Simu1_mRNA <- rbind(cbind(matrix(rnorm(40*5, mu[1], 0.1),40,5),
                          matrix(rnorm(40*5, mu[2], 0.1),40,5),
                          matrix(rnorm(40*5, mu[3], 0.1),40,5)),
                    matrix(rnorm(10*15,0,0.1),10,15))

mu <- c(1,3,-3)
Simu1_methyl <- rbind(cbind(matrix(rnorm(40*5, mu[1], 0.1),40,5),
                            matrix(rnorm(40*5, mu[2], 0.1),40,5),
                            matrix(rnorm(40*5, mu[3], 0.1),40,5)),
                      matrix(rnorm(10*15,0,0.1),10,15))

## feature modules across two datasets
group <- list(c(1:10,51:60), c(11:20,61:70), c(21:30,71:80), c(31:40,81:90))

DList <- rbind(Simu1_mRNA, Simu1_methyl)
dim(DList)

d <- t(DList)

gapResult <- gapStatistics(d,K=3,B=10,group=group)
print(gapResult)
}
\author{
Caleb
}
