% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISKmeans.R
\name{ISKmeans}
\alias{ISKmeans}
\title{IS-Kmeans}
\usage{
ISKmeans(d, K = NULL, gamma = NULL, alpha = 0.5, group = NULL,
  nstart = 20, wsPre = NULL, penaltyInfo = NULL, sparseStart = TRUE,
  silent = FALSE, maxiter = 20)
}
\arguments{
\item{d}{combined data matrix n*J, where n is number of subjects, J=J1+J2+... and J1 is number of features in omics dataset 1 and J2 is number of features in omics dataset 2...}

\item{K}{number of clusters}

\item{gamma}{Penalty on total number of features. Larger gamma will yeild small number of selected features.}

\item{alpha}{balance between group sparsity and individual sparsity. alpha=1 yeilds no group sparsity. alpha=0 yeilds no individual penalty.}

\item{group}{Prior group information. Potentially these group can contain overlap features. group is a list and each element of the list is feature index.}

\item{nstart}{Number of initials for Kmeans for sparse Kmeans}

\item{wsPre}{Initial feature weight.}

\item{penaltyInfo}{only for the purpose of gap statitics. Here we will fix the penalty design to perform gap statistics. The input should be a list of groupInfo. See groupInfo for details.}

\item{sparseStart}{Use Sparse Kmeans to do initialization.}

\item{silent}{Output progress.}

\item{maxiter}{Maximum numbre of iteration between ws and Cs.}
}
\value{
m lists, m is length of gamma parameters. Each list is consisting of
(ws=ws, Cs=Cs, objective=ADMMobject$objective, BIC=BIC, gamma=agamma,alpha=alpha
\item{ws}{weight for each feature. Zero weight means the feature is not selected.}
\item{Cs}{Cluster Assignment}
\item{objective}{objective value}
\item{obj0}{sum of weighted separation ability. This term is for the purpose of gap statistics.}
\item{groupInfo}{a list containing group design, alpha, gamma}
}
\description{
Integrative Sparse KMeans
}
\details{
Integrative Sparse KMeans, integrating multiple omics dataset, using prior group information.
}
\examples{
set.seed(123)

# Generate two random omics datasets
mu <- c(-3,1,3)
Simu1_mRNA <- rbind(cbind(matrix(rnorm(40*5, mu[1], 0.1),40,5),
                          matrix(rnorm(40*5, mu[2], 0.1),40,5),
                          matrix(rnorm(40*5, mu[3], 0.1),40,5)),
                    matrix(rnorm(10*15,0,0.1),10,15))

mu <- c(1,3,-3)
Simu1_methyl <- rbind(cbind(matrix(rnorm(40*5, mu[1], 0.1),40,5),
                            matrix(rnorm(40*5, mu[2], 0.1),40,5),
                            matrix(rnorm(40*5, mu[3], 0.1),40,5)),
                      matrix(rnorm(10*15,0,0.1),10,15))

## feature modules across two datasets
group <- list(c(1:10,51:60), c(11:20,61:70), c(21:30,71:80), c(31:40,81:90))

DList <- rbind(Simu1_mRNA, Simu1_methyl)
dim(DList)

## colSums(module)
K=3
nstart=20
silent=FALSE
maxiter=6
centers=NULL
error=1e-4
sparseStart = TRUE
gamma = 0.2
alpha = 0.5
d <- t(DList)

iRes <- ISKmeans(d, K=3, gamma=0.5, alpha=0.5, group=group)

}
\author{
Caleb
}
